requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
cassandra-driver==3.25.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Flask==2.2.3
geomet==0.2.1.post1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
prompt-toolkit==3.0.38
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
